requirements.txt
Django==2.2.1
ImageHash==4.0
numpy==1.16.3
Pillow==6.0.0
pytz==2019.1
PyWavelets==1.0.3
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
